Moved logical command classes to a separate file
